update order algo, refactor, rename, transf env
The ordering algorithm for the second stage of the ranking update in the
while loop was updated to be hierarchical (on two levels), rather than
the previous sum of the previous and updated rankings. is_prime was
moved to the Inspector module for neatness. An environment finding
algorithm for transformations was added, and files were renamed to
reflect their contents.